Count number of items in the project to determine if it is worth saving
We check that they have added a comment, state, or transitions since I felt those would be the things that someone would add
